requirements.txt
pyyaml>=6.0
jinja2>=3.1
pydantic>=2.0
